Basic implementation of view components.
ViewComponents are used to describe the structure of a rendered view
and sets it into relation with the model.

E.g., a conventional component is a NodeComponent. This tells that
a view component is a full fledged, top-level node view.
On the other side, a PropertyComponent is responsible for a node's property.